Refactor v2 lookup evaluation helpers
